Keep the encrypted unarchive, leave the rest of the branch behind
Reading encrypted archives is the part of this branch that is still needed and
still works. Writing them was already possible, so the library could produce
archives it could not read back. The password is now by name in the unarchiver
too, matching the archiver, so a secret need not be held for the lifetime of the
instance.

The tests of the size check and of reading a .tar.gz are kept as well. The first
covers Archiver.checkUncompressedSize directly, which until now was only reached
through the round trip suites, and it runs on Scala.js too. The second covers
composing an unarchiver with a decompressor, which the README describes and
nothing exercised.

Left behind: the suite renames, which now collide with the shared suites the
cancellation and unarchive work introduced, and the eager loading of the brotli4j
native library. When this branch was written only the compressor called
ensureAvailability; both paths do now, and the proposed fix ran as a side effect
of class initialisation, where the effect type cannot sequence it.

There is no round trip test for encryption on purpose. Zip4JArchiver takes a
password and never sets encryptFiles or an encryption method on the
ZipParameters, so it writes the entries in the clear and the password does
nothing. A round trip would pass either way.
