+++ brotli4j/src/main/scala/de/lhns/fs2/compress/Brotli4J.scala
@@ -17,9 +17,6 @@ class Brotli4JCompressor[F[_]: Async] private (chunkSize: Int, params: Encoder.P
 }
 
 object Brotli4JCompressor {
-  // Trigger loading the native library.
-  new Brotli4jLoader()
-
   def apply[F[_]](implicit instance: Brotli4JCompressor[F]): Brotli4JCompressor[F] = instance
 
   def make[F[_]: Async](
@@ -44,9 +41,6 @@ class Brotli4JDecompressor[F[_]: Async] private (chunkSize: Int) extends Decompr
 }
 
 object Brotli4JDecompressor {
-  // Trigger loading the native library.
-  new Brotli4jLoader()
-
   // Defined as DEFAULT_BUFFER_SIZE in BrotliInputStream, but isn't public
   def defaultChunkSize: Int = 16384
