+++ brotli4j/src/test/scala/de/lhns/fs2/compress/Brotli4JRoundtripSuite.scala
@@ -5,26 +5,11 @@ import cats.effect.std.Random
 import fs2.{Chunk, Stream}
 import munit.CatsEffectSuite
 
-import java.nio.charset.StandardCharsets.UTF_8
 import java.util
-import java.util.Base64
 
-class Brotli4JSuite extends CatsEffectSuite {
+class Brotli4JRoundtripSuite extends CatsEffectSuite {
   implicit val brotliCompressor: Brotli4JCompressor[IO] = Brotli4JCompressor.make()
   implicit val brotliDecompressor: Brotli4JDecompressor[IO] = Brotli4JDecompressor.make()
-
-  test("brotli decompress") {
-    val clear = Chunk.array("hello world!".getBytes(UTF_8))
-    val compressed = Chunk.array(Base64.getDecoder.decode("iwWAaGVsbG8gd29ybGQhAw=="))
-    for {
-      obtained <- Stream
-        .chunk(compressed)
-        .through(Brotli4JDecompressor[IO].decompress)
-        .chunkAll
-        .compile
-        .lastOrError
-    } yield assert(clear == obtained)
-  }
 
   test("brotli round trip") {
     for {
